Title page (#27)
This adds some styles for the main menu, settings and loading pages. It
also locks the entire game to a 16:9 aspect ratio. This means that we
don't have to worry about responsive design but by tying styles to
window width we can still make the most of the real estate we have.