Change how we do asset preloading and add a phone to typing game (#33)
Remove the "preload" function and instead load all minigame assets in
the background as per https://pixijs.com/8.x/examples/assets/background

This means by the time we call `Assets.load` in `populateContainer`,
hopefully the asset will have loaded in the background and will be
cached. However we can't guarantee this to be the case, so we need to
turn all our `populateContainer` definitions into async functions.