Up resolution and add timer progress (#34)
Adds in a little timer to all games. By default this rendered at like 6
pixels so updated pixi to make use of as much screen resolution as it
can.

This broke the rendering of the phone screen which was relying on no
resolution multiplier so manually set it's size to fix.